+++ src/main/java/ru/job4j/dreamjob/repository/Sql2oVacancyRepository.java
@@ -38,6 +38,9 @@ public class Sql2oVacancyRepository implements VacancyRepository {
 
     @Override
     public boolean deleteById(int id) {
+        if (findById(id).isEmpty()) {
+            return false;
+        }
         try (var connection = sql2o.open()) {
             var query = connection.createQuery("DELETE FROM vacancies WHERE id = :id");
             query.addParameter("id", id);
